Fix turning on org-journal-mode with custom org-journal-dir
When the `org-journal-dir` variable is set to some things, the default
`auto-mode-alist` for turning on `org-journal-mode` doesn't get
triggered. Figured that a unique 8 character filename with just
numbers would be unique enough, and shouldn't have to be in the
`journal` directory.

To ease in customizing the view of the journal entries, all formatting
for the `org-journal-date-format` and `org-journal-time-format` should
be in the variables.